Fix telegram_id to BigInteger for PostgreSQL compatibility

+++ backend/app/models/chat_history.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DateTime
 from datetime import datetime, timezone
 from app.database.database import Base
 
@@ -7,7 +7,7 @@ class ChatHistory(Base):
     __tablename__ = "chat_history"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, index=True, nullable=False)
+    telegram_id = Column(BigInteger, index=True, nullable=False)
     role = Column(String, nullable=False)  # "user" or "assistant"
     content = Column(Text, nullable=False)
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
